pyproject.toml
[tool.poetry]
name = "poc"
version = "0.1.0"
description = ""
authors = ["Abdullah Chaudhry <appleconda@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
python-dotenv = "^1.0.1"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
